+++ gradlePlugin/helpers-dependency/build.gradle.kts
@@ -20,7 +20,7 @@ plugins {
 }
 
 // To make it available as direct dependency
-group = "care.data4life.sdk.helpers.dependency"
+group = "care.data4life.gradle.fhir.helper.dependency"
 version = "1.0.0-SNAPSHOT"
 
 repositories {
@@ -28,8 +28,8 @@ repositories {
 }
 
 gradlePlugin {
-    plugins.register("care.data4life.sdk.helpers.dependency") {
-        id = "care.data4life.sdk.helpers.dependency"
-        implementationClass = "care.data4life.sdk.helpers.dependency.DependencyPlugin"
+    plugins.register("care.data4life.gradle.fhir.helper.dependency") {
+        id = "care.data4life.gradle.fhir.helper.dependency"
+        implementationClass = "care.data4life.gradle.fhir.helper.dependency.DependencyPlugin"
     }
 }
