+++ gradlePlugin/helpers-dependency/src/main/kotlin/care/data4life/gradle/fhir/helper/dependency/Version.kt
@@ -13,7 +13,7 @@
  * applications and/or if you’d like to contribute to the development of the SDK, please
  * contact D4L by email to help@data4life.care.
  */
-package care.data4life.sdk.helpers.dependency
+package care.data4life.gradle.fhir.helper.dependency
 
 object Version {
     object GradlePlugin {
